requirements.txt
asgiref==3.3.1
Django==3.1.7
django-cors-headers==3.7.0
django-filter==2.4.0
django-model-utils==4.1.1
djangorestframework==3.12.2
pytz==2021.1
sqlparse==0.4.1
